+++ brkt_cli/encrypt_ami.py
@@ -127,54 +127,6 @@ class InstanceError(BracketError):
     pass
 
 
-def setup_encrypt_ami_args(parser):
-    parser.add_argument(
-        'ami',
-        metavar='AMI_ID',
-        help='The AMI that will be encrypted'
-    )
-    parser.add_argument(
-        '--encrypted-ami-name',
-        metavar='NAME',
-        dest='encrypted_ami_name',
-        help='Specify the name of the generated encrypted AMI',
-        required=False
-    )
-    parser.add_argument(
-        '--encryptor-ami',
-        metavar='ID',
-        dest='encryptor_ami',
-        help='Bracket Encryptor AMI',
-        required=False
-    )
-    parser.add_argument(
-        '--key',
-        metavar='NAME',
-        help='EC2 SSH Key Pair name',
-        dest='key_name',
-        required=True
-    )
-    parser.add_argument(
-        '--validate-ami',
-        dest='no_validate_ami',
-        action='store_true',
-        help="Validate AMI properties (default)"
-    )
-    parser.add_argument(
-        '--no-validate-ami',
-        dest='no_validate_ami',
-        action='store_false',
-        help="Don't validate AMI properties"
-    )
-    parser.add_argument(
-        '--region',
-        metavar='NAME',
-        help='AWS region (e.g. us-west-2)',
-        dest='region',
-        required=True
-    )
-
-
 def get_default_tags(session_id, encryptor_ami):
     default_tags = {
         TAG_ENCRYPTOR: True,
@@ -322,7 +274,7 @@ def _append_suffix(name, suffix, max_length=None):
     return name + suffix
 
 
-def _get_encryptor_ami(region):
+def get_encryptor_ami(region):
     bracket_env = os.getenv('BRACKET_ENVIRONMENT',
                             BRACKET_ENVIRONMENT)
     if not bracket_env:
